+++ tests/e2e/Services/Account/AccountBase.php
@@ -365,7 +365,7 @@ trait AccountBase
 
         $this->assertContains($response['body']['logs'][2]['event'], ["users.{$userId}.create", "users.{$userId}.sessions.{$sessionId}.create"]);
         $this->assertEquals($response['body']['logs'][2]['ip'], filter_var($response['body']['logs'][2]['ip'], FILTER_VALIDATE_IP));
-        $this->assertEquals(true, DateTime::isValid($response['body']['logs'][2]['time']));
+        $this->assertEquals(true, DateTime::isValid($response['body']['logs'][2]['duration']));
 
         $this->assertEquals('Windows', $response['body']['logs'][2]['osName']);
         $this->assertEquals('WIN', $response['body']['logs'][2]['osCode']);